Fix job.dependencies_are_met() if dependency is canceled
