+++ rq/job.py
@@ -1002,12 +1002,12 @@ class Job:
                             for _id in connection.smembers(self.dependencies_key)}
 
         if parent_job:
-            # If parent job is canceled, no need to check for status
+            # If parent job is canceled, treat dependency as failed
             # If parent job is not finished, we should only continue
             # if this job allows parent job to fail
             dependencies_ids.discard(parent_job.id)
             if parent_job._status == JobStatus.CANCELED:
-                pass
+                return False
             elif parent_job._status == JobStatus.FAILED and not self.allow_dependency_failures:
                 return False
